refac: 기본시급 인풋 컴포넌트화
