+++ src/features/Create-Shop/CreateShop.tsx
@@ -11,11 +11,11 @@ import Input from '@/shared/ui/Input';
 import Text from '@/shared/ui/Text';
 import useCreateShopRequest from './hooks/useCreateShopRequest.tsx';
 import { generatePresignedUrl, uploadImageToS3 } from './model/Api.ts';
-import { MINIMUM_WAGE } from '@/shared/constants/Wage.ts';
 import ShopNameInput from './ShopNameInput.tsx';
 import CategorySelect from './CategorySelect.tsx';
 import AddressSelect from './AddressSelect.tsx';
 import AddressDetailSelect from './AddressDetailSelect.tsx';
+import OriginalHourlyPayInput from './OriginalHourlyPayInput.tsx';
 
 const CreateShop = ({ initialValues, shopId }: CreateShopProps) => {
   const {
@@ -51,23 +51,6 @@ const CreateShop = ({ initialValues, shopId }: CreateShopProps) => {
     generatePresignedUrl({ imageName, setPresignedUrl });
     setValue('imageUrl', '');
     setUploadedImageUrl('');
-  };
-
-  const handlePayChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const hourlyPay = Number(e.target.value.replaceAll(',', ''));
-    // eslint-disable-next-line no-restricted-globals
-    if (isNaN(hourlyPay) || !hourlyPay) {
-      setValue('originalHourlyPay', '');
-    } else {
-      setValue('originalHourlyPay', hourlyPay);
-    }
-  };
-
-  const handlePayBlur = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const pay = Number(e.target.value);
-    if (pay) {
-      setValue('originalHourlyPay', pay.toLocaleString('ko-KR'));
-    }
   };
 
   useEffect(() => {
@@ -107,42 +90,19 @@ const CreateShop = ({ initialValues, shopId }: CreateShopProps) => {
           />
         </div>
         <div className='flex flex-col items-center gap-5 lg:flex-row'>
-          {/* 주소 */}
           <AddressSelect
             control={control}
             initialValues={initialValues}
             errors={errors}
           />
-          {/* 상세 주소 */}
           <AddressDetailSelect register={register} errors={errors} />
         </div>
         <div className='flex flex-col items-center gap-7 lg:items-start'>
-          {/* 시급 작성  */}
-          <div className='relative flex w-[350px] flex-shrink-0 flex-col items-start gap-2 md:w-[472px]'>
-            <label htmlFor='originalHourlyPay'>기본 시급 *</label>
-            <Input
-              width='w-full'
-              id='originalHourlyPay'
-              type='text'
-              placeholder='입력'
-              {...register('originalHourlyPay', {
-                required: '시급을 입력해주세요.',
-                onChange: e => handlePayChange(e),
-                onBlur: e => handlePayBlur(e),
-                validate: {
-                  checkWage: (v: string) =>
-                    Number(v.replaceAll(',', '')) >= MINIMUM_WAGE ||
-                    '최저임금 보다 적습니다.',
-                },
-              })}
-            />
-            <span className='absolute right-4 top-[51px]'>원</span>
-            {errors.originalHourlyPay && (
-              <span className='absolute left-1 top-[105px] text-xs text-red-500'>
-                {errors.originalHourlyPay.message?.toString()}
-              </span>
-            )}
-          </div>
+          <OriginalHourlyPayInput
+            register={register}
+            errors={errors}
+            setValue={setValue}
+          />
           {/* 가게 이미지 추가하기  */}
           <div className='relative flex h-[400px] w-[350px] flex-col gap-2 overflow-hidden md:w-[472px]'>
             <div>가게 이미지 *</div>
